Add unused doors counter
The counter keeps trach of the unused doors. This information is
used in the map generation process to eventually add new doors and
avoid dead ends before having reached a flight of stairs to the
next floor.

This commit also restores a few utils methos to render the map
via CUI.